Refactor: Cleanup and organize add-on related imports
